Remove default CMake toolchain files
This turned out to be not so great idea as there are packages (namely flatbuffers) which try to detect the toolchain files and disable all their options. Diagnosis requires reading CMakeLists.txt for the package. In case there are other such packages in the wild it is better not to have defaults which enable surprising behaviour.