Try harder to call git with correct options
So many years I'm doing this and still it is surprisingly hard to implement
the "update" action to do the "right thing". I keep finding cases when things
break subtly or not so subtly especially if we want to leave the repository in
the state usable for a normal interactive Git usage.

This is another attempt. Hopefully it makes the goal of having a robust Git
wrapper closer not farther. Corollary, the 'update' command should start
working again.