Improve robustness of git source properties handling
Test and fix a few corner cases related to various combinations of
source properties for type GitSource. Allow disabling "origin", though
the resulting commands might not be very useful for now but at least it
does not cause assertion failures. Also remove special treatment of
branch name "master". This facilitates a better interoperability with
manual git usage and allows to have a branch in the repository which
does not need to be specified in rules. This, for example, allows to
switch branch in layer repo and update should work as expected.