Modify the default output mode of build command to not use tail
Tailing the output by default intermixes the outputs for independent
targets which can be confusing and also there is a problem when cmd
wrapper kills tail too early and the last lines from the log will not be
printed on the console which can be even more confusing. These problems
are especially apparent when running under the Windows Subsystem for
Linux where the console is very slow.

The new mode lets the Ninja buffer the whole command output and display
it only after the command completes just as it normally does. This mode
also does not need tail. The output of all commands is still saved to
logs.

The previous behaviour (with tail) is still useful for long compiling
packages and such and can be enabled with '--follow' argument for build.