Make build output less verbose by default
After a few days of testing it became apparent that showing the output
of all targets by default is a bit extreme. Some build systems write a
lot and commands run in parallel, so it becomes hard to find the errors
as the console scrolls up rapidly.

A new default is almost the opposite: do not print the output at all
unless the last command in the stage failed. This also works in the
"follow" mode (which was the default for a long time), i.e. the log of a
target which failed will be shown even if it was not specified in the
command line to follow but was built to satisfy a dependency.