Fix exit hook of jagen-stage overriding exit status
If the stage exiting for a reason different from pkg_run returning non
zero status (for example run_stage failing) the exit hook was overriding
exit status with the status of last pkg_run command which might be 0.
This caused the stage exit status to not be propagated to the caller and
build continuing despite the failure.