Make die set last_error too
When a build exits by executing the 'die' function instead of a failed
pkg_run the error log is not shown on the console automatically. This is
because only pkg_run is setting global error status which is used by the
exit hook, now die sets it too.

Probably we still need to wrap all other calls to exit to make sure the
log is always shown on error.