Fix init to clone Jagen into the correct location when piped to Shell
Apparently the init does the things in an incorrect order since the '-d' option
was added. Looks like I've not tested it thoroughly enough. Such a blunder.

Also changed it check for non-empty directory before trying to clone and ignore
hidden files when doing so which was making the usage unnecessary inconvenient.