Add ccache flag on init if ccache command was found
I was using Jagen with ccache for years and have not had any problems so far.
It's hard to think of a situation when you might not want to use it apart from
some kind of one-off builds. Normal use case for a build system to is to build
software, probably many times and ccache speeds up this process significantly.

It is possible to suppress the default behaviour by passing -ccache flag to
init, e.g.: `.../init -a -ccache`.