Make clean and update actions more consistent
There were differences in the behaviour depending on the way how the clean or
update were called: from a build stage or directly by source subcommand. More
things to remember and opportunities to shoot yourself in the foot.

Now the stage just calls the src subcommand and all the logic is implemented in
the same place. Also differences in handling excludes and dirty status between
clean and update were also removed.