Add options to ignore dirty and exclude
Source clean, update and build commands now have options '--ignore-dirty' and
'--ignore-exclude' which cause them to ignore dirty or exclude state of the
source directories respectively.

Ignoring dirty state is especially useful for cases when you have project or
backup or some other files lying around the repositories and just want to build
it quickly without removing them or messing with ignores.

Ignoring exclude is useful for one-off cleaning or updating without adjusting
the rules back and forth just for the task which can be annoying.