Run clean stages in order
After adding source directory to clean list I've found that in the case of
in_source build packages git complains about index.lock because multiple
processes are trying to access the same source directory (one for shared target
and another for config-specific). I've thought about some heuristics to not run
it maybe in certain cases but this just makes the behaviour harder to predict
and might have surprising interaction with custom clean directories as well. So
just ordering clean stages one after another solves the problem reliably. Seems
that an overhead for running git clean on already clean directory is not that
big.