Skip env in root dir on import
When the root dir itself is added to layers this leads to infinite recursion
because env.sh in root_dir sources global env which tries to import env again.
The refresh just hangs. Regardless of whether it makes sense or not to have such
a configuration looping indefinitely here is not useful for anything.