Remove explicit "generate" stage
Move the relevant steps to the "patch" stage instead.

It is very rare to have a need to regenerate the build system on already already
unpacked packages and an existence on special stage which appear only with
certain properties makes the build system harder to explain without much
practical value.