New default build output mode which supports colors
If neither "follow" nor "print" options are set add explicitly specified build
targets to "console" Ninja pool which lets them detect if their output stream is
connected to the terminal and display colors if they are able. Consequently
those targets are always executed sequentially and their output is not captured
into Jagen's logs. The other dependent targets are executed in background and
their output is saved as before.

This is especially useful for Rust which displays nice colored error messages
and also for more modern C++ compilers.

This commit also removes "--interactive" options from the build command
as it became useless.