Do not consider implicitly added "clean" stages as interactive targets
When "--clean" option is given for the build and no pattern in the
arguments match "clean" stages those are added automatically to the
list. We do not consider them "interactive" and do not add them to the
"console" pool. Those are not supposed to display anything normally
anyway and this makes build output slightly neater.