Implement auto-defining library packages for build
If build argument specify a package which was not explicitly defined in rules
but found in libraries (pkg/*.lua files in layers) Jagen will define it
automatically for one-shot build.

This feature can be used for "extra" packages built on demand or for testing
library packages from layers without requiring adding them to rules manually one
by one.