Fix refresh run twice with auto packages
When .auto-packages changes it causes refresh, then refresh is run once
more claiming that the recorded mtime of one of the includes is earlier
than the most recent input .auto-packages. I could not figure out why
though: includes are regenerated on the first refresh as well, seems
that this is not recorded when ninja reload due to build.ninja changing
as well.

Making includes into phony inputs of build.ninja instead of outputs of
the refresh rule solves this problem.