Improve generator interaction with interactive targets
When the target is specified on the command line its pool is changed to
'console'. This will enable the targets having assigned pools such as
gradle-anrdoid and rust-toolchain to be able to run simultaneously with the
interactive ones which defeats the purpose of having assigned pools in the first
place. To avoid this, now if one of the assigned targets is interactive the
others from the same pool will become interactive too.