Do not set LUA_PATH in the project env globally
I've got a nasty surprise on MacOS with case-insensitive filesystem when
the rules.lua file in the project was found before Rules.lua in the
source code and broke everything. Now LUA_PATH contains only the path to
the Jagen's src directory and is set for the command itself instead of
the whole environment to be extra safe.

So, Jagen on MacOS should work again.

This change will break usage of 'require' in rules to include other
modules from jagen_path which was allowed previously. Maybe I will
reintroduce this functionality later.