Show attempted paths when patch or supplementary file was not found
Apart from better diagnostics which is nice also rewrote the logic in
pure Lua to avoid forking the shell just to look for some files. This is
like 10000 times faster (according to tests) but unlikely to be
noticeable due to usually small number of those lookups which sums up to
some milliseconds in total.