Consider also install only packages when defining default config
Previously automatic host config define was added if the package has non-empty
build type. Sometimes it is useful to define virtual packages which do not have
to build anything but pull others by require. Currently specifying install =
true is require for such packages. I will consider to remove this requirement as
well if possible with the current rule handling logic.