Do not make shallow clones of Git repos if rev is set
Unless the rev is the same as HEAD the shallow clone will not have it and the
checkout will fail. The fix is to fetch full history if the rev is set. For now
there is no handling for the case if the repo was cloned shallowly and the rev
was set later. To do some other time.