Improved project mode support in init script
Work in progress to make a "project mode" feature actually useful.

In contrast with the traditional "workspace mode" where there is no
central repository and the build root itself contains all sources and
artifacts, in the "workspace mode" the root is created inside of the
"main project" source directory to manage its dependencies with Jagen.
This is also intended to make the setup for CI more straightforward.