Adjust 'source clean' to delete ignored files for non-excluded sources
Also make the cleanup log message more informative.

This partially reverts the decision introduced in 38ed13da (2019-02-21)
to not delete ignored files by default. Now these files will be cleaned
up again but only for the packages which are not manually excluded.
This is more convenient for dependencies pulled as sources which like to
pollute their source directories with generated files and such.

In contrast, the packages specifically added to jagen_source_exclude or
having source.exclude set are considered being actively worked on. Doing
just a standard clean is more common for them to not wipe IDE configs
and state files or tag indexes. An additional '--ignore' option is
required to clean all when ignoring the exclusion.