Set the pkg_install_prefix to empty in target config
Apparently the CMake started to recognize the pathnames starting with
the double slash as network paths and complain during the installation
that this is not supported. The double slash might appear after the
expansions if the CMAKE_INSTALL_PREFIX is referenced in the CMakeLists
directly.

Setting it to empty for now but there was probably a good reason for it
to be '/' historically. Probably build breakages will follow after this.
With fixes to follow. It will take time to test all use cases.