Set CMAKE_SYSTEM_PROCESSOR when cross-compiling
Though rare, some packages have cpu-specific branches in their
CMakeLists which might fail if the CMAKE_SYSTEM_PROCESSOR is empty which
is the default when cross-compiling.

With this change the CMAKE_SYSTEM_PROCESSOR will be set in the toolchain
file from pkg_build_cpu (build.cpu) if it is set or from pkg_build_arch
(build.arch) as a fallback which itself is determined from the first
component of the build.system property. This property is inherited from
the toolchain, so no manual changes to the rules should be necessary.