Adjust table merging semantics with false/empty values
When a table key value was merged to non-table type the non-table value was
converted to a single-value table and the merge recursed into the resulting 2
tables. This caused an undesired effect in the case when the destination value
was explicitly 'false'. This 'false' value became the first member of the
resulting list.

Now the single-value table conversion will be performed only for the values
which evaluate to 'true', so the non-existing keys of an explicit 'false' will
have the same effect as if the destination value was an empty table which will
effectively be replaced by the source table during the merge.

Additionally the special merging semantics now apply only to non-empty source
key values which means that the existing non-empty table value can be reset to
empty by assigning an empty table value.

Important to note that this might have an unexpected effect when editing rules:

....
build = {
  -- options = { 'a', 'b' }
}
....

will set all the build keys to empty because with the commented out options this
statement is equivalent of:

....
build = {}
....

previously the empty table values had not effect.

This new treatment of empty table values is experimental. Backward compatible
way to achieve the same effect is to set the key to 'false'.