Adjust handling of use specs in jagen-stage
The use specs of the form "name:config" are supposed to override the
config of the currently executing stage. There was a confusion in the
export filename expansion handling which caused a config-specific export
file to be included twice when executing a stage with a config.

Now, for the "name:config" use specs, a generic export file is included
first (name.export.sh), then the config-specific one
(name:config.export.sh). This will work the same way regardless of where
the use spec is coming from or which stage is executing: either the
generic one or the config-specific.

For bare "name" use specs:

 - When executing a generic stage only generic name.export will be
   included.
 - When executing a config-specific stage the generic name.export will
   be included first, then config-specific one with the same config as
   the currently executing stage.