Rewrite rule engine to a template-based
The rule engine is fully rewritten to use a cascading template rules
structurally applied to the packages matching the patterns. It was
resembling that internally already but with the rules explicitly coded
and many cases added over the years. Not all the cases were added back
yet and I've decided to use the opportunity to loose some to make the
implementation more straightforward.

This is a major breaking change. All the differences are too many to
count and I've not yet decided which to retain so this is to be
documented sometime later before the major release increase.

Additionally to the new template rules the 'use' directive was added
which declares module dependencies and makes the Jagen engine to load
the specified rule file and add it's path to the 'layer path'. This
replaces the 'jagen_layers' definition from the 'config.sh'. The
'package' files are now just another module file which can have
arbitrary rules and use directives instead of defining kind of special
generic package rule.

The concept of a mandatory package config (either 'host' or 'target')
was mostly removed from the code.

The Jagen does not try to define package rules automatically anymore.
All uses should be defined by an explicit rule either in the same module
or loaded by a dependency lookup.

The package 'requires' were removed as the 'uses' does about the same
thing. All the use imports and exports are resolved at the rule
generation time. The packages 'export' stage was removed. The shell code
with all the evals at the build time was removed as well. The rules can
still refer to their other parts with shell-like variable references
which are resolved by the engine during the loading.

The rule validation code was removed for now due to large changes to the
internal structures. Right now the rule error handling is rudimentary.
The validation will need to be implemented later.